Auto-generates hyperlinks from file system folder
This one takes a folder in the file system, reads its files, stores their path in a MySQL database, then calls the DB info to auto-generate an index-like page or a very basic image gallery with image thumbnails that are also links to the full sized image (Very basic).  Built in browser security settings block this though, so the gallery is only tested to work on Internet Explorer, while Firefox specifically only displays the listing.